Fix "permission denied" bug for multi-user systems
On a multi-user system Eclipse might report a "permission denied"
message for a pom.xml file. This is known to be caused by
a workaround in mavenarchiver. For more details, see
https://dev.eclipse.org/mhonarc/lists/m2e-users/msg04160.html

mavenarchiver creates a directory "fakerepo/" in ${"java.io.tmpdir"}
with 755 permissions in Linux. As a result, other users will not be
able to write to this directory.

This changset fixes the issue by assigning a unique fake repo
directory for a user based on ${"user.name"}.